Bullets, lives, waves and misc fixes
When a bullet hits an enemy or a player, they lose a life. If an enemy's life drops below 0, it is removed from the Vector of enemies in the scene, if the player's lives drop below 0, the game is ended. Static enemies fire bullets to the bottom of the screen at random. Implemented waves that increase in difficulty through number of enemies, number of lives those enemies have and the chance that they fire a bullet in a given update loop. Fixed bug where the top bar would return if you minimised and then reopened the game. Added delta time constraints to the game loop, much more responsive. Finally, showed the lives and wave number on screen and commented and cleaned up all of the code.